For version 2.1.1 resolving issues: ES2-95 ES2-91 ES2-92 ES2-86 ES2-75 ES2-68

+++ src/apps/productmanagement/requests.py
@@ -109,51 +109,60 @@ def create_archive_from_request(request_file):
         return 1
 
     my_product = my_request['product']
-    my_mapsets = my_request['productmapsets']
     my_version = my_request['version']
+    my_mapsets = []
+    if any(x == "productmapsets" for x in my_request):       # hasattr(my_request, 'productmapsets')
+        my_mapsets = my_request['productmapsets']
 
     n_mapsets = len(my_mapsets)
-    incresing_number=1
+    incresing_number = 1
 
-    # Loop over defined mapsets
-    for my_mapset in my_mapsets:
-        mapsetcode = my_mapset['mapsetcode']
+    if n_mapsets > 0:
+        # Loop over defined mapsets
+        for my_mapset in my_mapsets:
+            mapsetcode = my_mapset['mapsetcode']
 
-        mapsetdatasets = my_mapset['mapsetdatasets']
+            mapsetdatasets = my_mapset['mapsetdatasets']
 
-        # Loop over all datasets in a mapset
-        for mapsetdataset in mapsetdatasets:
-            subproductcode =  mapsetdataset['subproductcode']
-            missing_info = mapsetdataset['missing']
-            archive_base_name=request_file.replace('.req','')
-            archive_name=archive_base_name+'_{0:04d}'.format(incresing_number)+'.tgz'
-            self_extracting_name=archive_name
-            self_extracting_name=self_extracting_name.replace('.tgz','.bsx')
-            logger.debug( 'Archive file name: {0}'.format(archive_name))
+            # Loop over all datasets in a mapset
+            for mapsetdataset in mapsetdatasets:
+                subproductcode = mapsetdataset['subproductcode']
+                missing_info = mapsetdataset['missing']
+                archive_base_name = request_file.replace('.req', '')
+                archive_name = archive_base_name+'_{0:04d}'.format(incresing_number)+'.tgz'
+                self_extracting_name = archive_name
+                self_extracting_name = self_extracting_name.replace('.tgz', '.bsx')
+                logger.debug('Archive file name: {0}'.format(archive_name))
 
-            # Create a product object - no date indication
-            product = Product(product_code=my_product, version=my_version)
-            [tarfile , results] = product.create_tar(missing_info, filetar=archive_name, tgz=True)
-            logger.debug('Files found for {0}: {1}'.format(subproductcode,results['n_file_copied']))
-            # Test there is - at list - 1 file missing
-            if results['n_file_copied'] > 0:
+                # Create a product object - no date indication
+                product = Product(product_code=my_product, version=my_version)
+                [tarfile , results] = product.create_tar(missing_info, filetar=archive_name, tgz=True)
+                logger.debug('Files found for {0}: {1}'.format(subproductcode, results['n_file_copied']))
+                # Test there is - at list - 1 file missing
+                if results['n_file_copied'] > 0:
 
-                logger.info('Creating file {0}'.format(self_extracting_name))
-                # Get the decompression script template
-                decompress_file = es_constants.decompress_script
+                    logger.info('Creating file {0}'.format(self_extracting_name))
+                    # Get the decompression script template
+                    decompress_file = es_constants.decompress_script
 
-                target = open(self_extracting_name,'wb')
-                shutil.copyfileobj(open(decompress_file,'rb'),target)
-                shutil.copyfileobj(open(archive_name,'rb'),target)
-                target.close()
-                os.chmod(self_extracting_name,0775)
-                # Increase the counter
-                incresing_number+=1
+                    target = open(self_extracting_name, 'wb')
+                    shutil.copyfileobj(open(decompress_file, 'rb'), target)
+                    shutil.copyfileobj(open(archive_name, 'rb'), target)
+                    target.close()
+                    os.chmod(self_extracting_name, 0775)
+                    # Increase the counter
+                    incresing_number += 1
 
-            # Remove .tgz file
-            os.remove(archive_name)
-            product = None
-    return
+                # Remove .tgz file
+                os.remove(archive_name)
+                product = None
+        else:
+            # Generate an archive with the complete datasets of all the subproducts.
+            # Which mapset? Get all mapsets of product from its Product instance and choose the Africa one?
+            x = 1
+
+    return 0
+
 
 def get_archive_name(productcode, version, id):
 
@@ -161,4 +170,71 @@ def get_archive_name(productcode, version, id):
     filename += 'archive_'+productcode+'_'+version+'_'+id+'.tgz'
     return filename
 
+
+def get_request_filename(productcode, version, subproductcode, mapsetcode, date=None):
+
+    filename = productcode + '_' + subproductcode + '_' + mapsetcode + '_' + version
+    if date is not None:
+        filename = productcode + '_' + subproductcode + '_' + mapsetcode + '_' + version + '_' + date
+
+    return filename
+
+
+def create_archive_vars(productcode, version, mapsetcode, subproductcode, from_date=None, to_date=None, time_suffix=None, output_dir=None):
+
+    # Creates an archive file (.tgz) for a single period (prod/version/sprod/mapset)
+    incresing_number=1
+
+    if output_dir is None:
+        output_dir = es_constants.es2globals['base_tmp_dir']
+
+    if time_suffix is not None:
+        time_token=str(time_suffix)
+    else:
+        if from_date is not None:
+            time_token=from_date
+            if to_date is not None:
+                time_token+='_to_'
+                time_token+=to_date
+        else:
+            time_token='alltimes'
+
+    # Define archive name
+    archive_base_name=output_dir+os.path.sep+ \
+                      productcode+'_'+version+'_'+mapsetcode+'_'+subproductcode+'_'+time_token
+
+    archive_name=archive_base_name+'_{0:04d}'.format(incresing_number)+'.tgz'
+    self_extracting_name=archive_name
+    self_extracting_name=self_extracting_name.replace('.tgz','.bsx')
+
+    logger.debug( 'Archive file name: {0}'.format(archive_name))
+
+    # Create a product object - no date indication
+    product = Product(product_code=productcode, version=version)
+    [tarfile , results] = product.create_tar_vars(productcode, version, subproductcode, mapsetcode, from_date=from_date,
+                                                  to_date=to_date, filetar=archive_name, tgz=True)
+
+    logger.info( 'Tar archive created: {0}'.format(archive_name))
+    logger.debug('Files found for {0}: {1}'.format(subproductcode,results['n_file_copied']))
+
+    # Test there is - at list - 1 file missing
+    if results['n_file_copied'] > 0:
+
+        logger.info('Creating file {0}'.format(self_extracting_name))
+        # Get the decompression script template
+        decompress_file = es_constants.decompress_script
+
+        target = open(self_extracting_name,'wb')
+        shutil.copyfileobj(open(decompress_file,'rb'),target)
+        shutil.copyfileobj(open(archive_name,'rb'),target)
+        target.close()
+        os.chmod(self_extracting_name,0775)
+        # Increase the counter
+        incresing_number+=1
+
+    # Remove .tgz file
+    os.remove(archive_name)
+    product = None
+
+    return self_extracting_name
